feat: ENT-7851 Added course section title for learners's first visit
