+++ src/components/dashboard/data/constants.js
@@ -1,1 +1,9 @@
 export const LICENSE_ACTIVATION_MESSAGE = 'Your license was successfully activated.';
+
+export const COURSE_SECTION_TITLES = {
+  current: 'My courses',
+  completed: 'Completed courses',
+  savedForLater: 'Saved for later',
+  assigned: 'Assigned Courses',
+  firstTimeUserAndAssigned: 'Your learning journey starts now!',
+};
